+++ http/mod.ts
@@ -100,6 +100,8 @@ export * from "./signed_cookie.ts";
 export * from "./server_sent_event_stream.ts";
 export * from "./user_agent.ts";
 export * from "./file_server.ts";
-export * from "./route.ts";
+// We keep this re-export as an exception for now as it's used in
+// `deno init --serve` output
+export * from "./unstable_route.ts";
 export * from "./header.ts";
 export * from "./method.ts";
